+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add handling of application/vnd.api+json media type.
 - Support passing models into query parameters (#316). Thanks @forest-benchling!
 - Add support for cookie parameters (#326).
+- New `--file-encoding` command line option (#330). Sets the encoding used when writing generated files (defaults to utf-8). Thanks @dongfangtianyu!
 
 ### Changes